Changes: API + Directory Structure
changed the directory structure for backend. API versioning enabled.
